Refactor all CCT opening tags, titles and contents
This fixes #84 and #86.

This makes all tags consistent, making titles
consistent (we previously used h1, h2 and h3 in random places) and
ensures that only contents and titles are displayed if we're (pretty)
sure there's content to be shown. We use Viper's default <p></p> value
to filter out areas that are empty.

This may need revisiting when the fix for #12 is installed.  We don't
know yet what that looks like.