Merge user-defined CCT classes & template classes
This fixes #83 by pulling out the classes that may be specified on a
Content Block and merging the required classes together.  Finally, we
output the result to the page.  This is two-pronged approach -- we open
the wrapper tag in the CCT header and then finish the tag (with the
class attribute) in the CCT Paint Layout itself (because we need to know
what template we're using).

If we had access to more info about *which* CCT we're rendering in the
header, we could be more generic.