SSJS: add DOM implementation via domino
This provides the ability to construct DOM objects, manipulate HTML and
more within the server-side JavaScript environment (SSJS). You can even
run jQuery on the server if that's your thing.

The cost of including the DOM + jQuery is ~0.2s (15%) for every raw page
render, but that's less of an issue once a page is cached.

This grants the choice of whether to use HTML manipulation code within
Matrix's keyword modifiers, or to do things within SSJS instead.