New Disable Link metadata item in Summary Card Card Details section
If Disable Link is selected then whenever this card is displayed in a
Summary Card List there will be no links back to the original asset nor
to any external URL that may have been specified. This was added because
sometimes you just want a card (e.g. for a student who doesn't have a
profile anywhere).