Image links can now specify new browsing content target if Link asset used
If the asset in question is a Link asset type and the user has specified
'external' in the link relation field then we will give the a tag a
target of _blank. We had to overload the relation field as the Link
asset doesn't have a target field. See Issue#107