Ensure Repeats line isn't visible on single events
This fixes #118 by checking if the event is a single event first.  The
problem was that the keyword replacement for %event_frequency% remains
verbatim on single events (when in other areas of Matrix it will get
replaced with null).  We handle this quirk by checking the type of the
event first.

Oddly, if you were to print() the value of `frequency` the result was
null/empty, because a later keyword replacement layer fixes the problem.

This is probably yet-another-Matrix-bug we need to report.