Handle "null" for titles/subtitles. Fixes #138
Previously, the title/subtitle renderer was failing to account for the
fact that the keyword for the sometimes-optional subtitle could have
been null.  This meant the renderer got a string literal of `null`
passed to it for the subtitle from the Matrix keyword as when the NULL
value in PHP is passed to `json_encode` it returns the string of "null".
Therefore, our code considered it to have a value.

However, when printing the keyword directly (eg without going through
`json_encode`) NULL in PHP represents nothing, so it wouldn't show up.

This handles the situation by using Matrix's `^empty:` modifier,
coercing the NULL into a empty string before being JSON-ified.

This also adds the same coercion to handle any future situation where
title might be optional but subtitle isn't.  Anyhoo...