Fix for the Summary Cards content block print layout
The URL for the link associated with the images was being treated
(visually) as part of the image and being cropped to a circle for
circular images. It looked terrible. I rearranged the CSS such that the
circular shape and flair were only applied to the image and not to the
link as well.