Summary Cards CB - Brighter integration and added new Tile style to the Grid and Mosaic layouts
There is a new metadata option in the List Settings called Card Style,
it defaults to Card (preexisting look) and has one other option; Tile.
The Tile look is image and asset title only with the title displayed
below the image in a dark blue footer block.
The Brighter integration consists of looking for the image in a specific
metadata field if it isn't available as a thumbnail (and the asset
itself isn't an image).