Adding Button Text override to Summary Card metadata schema
The Button Text for an individual Summary Card can now be customised.
This will override the Summary Cards list Button Text for any list it is
in. I've also enabled the external site link icon for the button links
in column layout lists; previously it was enabled for grid layouts only.