Make linking to images in Content Blocks optional
This fixes #143 by implementing a separate paint layout for non-linked
related asset images.  This is the only way around this problem because
we're relying on SSJS to print keywords (and we can't pass arguments
to a Paint Layout or SSJS block that doesn't yet exist).  So, we have
had to reproduce our Paint Layout with different functionality.