Navbar button and wrapper are hidden if no content
This fixes #144 by introducing a conditional that checks if the site
(the top-level navigation) has any corresponding TYPE_1 (in-navigation)
assets.  The only time the menu_normal Design Area and the "Immediate
Children" asset listing could differ is for asset types not listed in
the asset listing.  However, we cover all the standard types and their
parent classes for assets that might have been created via the Edit+
interface.

If there are other asset types that should be navigation but aren't
showing up, then add them into asset ID 379674.