Avoid function hoisting for Feature Number Cards
This fixes #145 by defining the function as a variable which can be
overwritten by each subsequent content block.  Previously, it was "last
Feature Number Cards" block wins because of hoisting.  Normally, this
sort of thing wouldn't be a problem but the asset ID was present in the
function definition, resulting in the final content block's description
being printed out in the wrong blocks.