Fix Edit+ integration for v5.5.3.0
This repairs the jQuery integration because the JCU Web Framework wants
a specific version of jQuery (>1 but <3) and Edit+ has a newer version
now.

This also fixes various graphical display issues for collapsing
headings, where the section headings now feature the descriptions
directly as part of the text.  This works around #147 for now.